+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ## Unreleased
 
 ### New
+- adding universal environment replace in manifests (ref: `${SOMEKEY}`)
 
 ### Changes
 - enforce strict validation for unknown values in manifests